make verify proof utility method private
